skills/vector-store/pyproject.toml
[project]
name = "vector-store-skill"
version = "0.1.0"
description = "Transient vector store service for fast similarity search"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "numpy",
    "faiss-cpu",
    "pydantic",
    "requests"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py"]
